mkDerivation: require a list for cmakeFlags
Currently, cmakeFlags can be passed as a string or a list.
Passing a string works OK for small cases but it becomes hard to manage
interspersing spaces between the flags with many options, and doesn't
match the semantics of each flag being a separate item.
Passing a list (of strings) is a better fit for the type of data, but
it currently relies on Nix's regular conversion to an environment
variable, which simply concatenates the contents with spaces.
This breaks if any of the flags themselves contain spaces.
cmakeFlagsArray was added as a way to get around this limitation,
but is unsatisfactory to use because items must be appended via bash
in a preConfigure hook.

So, update cmakeFlags to be more ergonomic:
  - Require passing a list of strings for cmakeFlags,
    and update all uses in nixpkgs.

  - Pass the list to the builder in a smarter way: instead of relying
    on Nix's conversion to a string, perform our own conversion by
    interspersing tabs.
    In case a literal tab is passed, it will be automatically escaped
    for the shell and behave properly.
    This also fixes passing flags that contain spaces.

  - Make the list available during preConfigure as a bash array, so any
    dynamic modifications to the CMake flags can be done there.

These changes make also cmakeFlagsArray redundant, so it has been
removed and replaced in all cases with cmakeFlags.

However, since `cmakeFlags` is processed by `mkDerivation`,
`overrideDerivation` should not be used to easily update `cmakeFlags`.
Using `overrideDerivation` and passing a list of strings will cause it
to be passed directly to the underlying derivation, bypassing the
processing in `mkDerivation`, and reverting back to the space-joining
behavior of Nix. Additionally, the original list of strings will not be
available for easy appending, only the processed string.
Instead, `overrideAttrs` should be used, which provides access to the
original `cmakeFlags`, and which will pass the resulting attrs to
`mkDerivation` so the processing can take place.