mkDerivation: use sets for cmakeFlags
Almost all flags passed to `cmakeFlags` are setting CMake cache options
with "-D<name>=<value>" or "-D<name>:<type>=<value>". Because the CMake
cache is essentially a set of keys and values, make it possible to use a
set for `cmakeFlags` as well (as an argument to `mkDerivation`).

`mkDerivation` will take an incoming `cmakeFlags` set and convert each
key/value pair to an appropriate "-D" option for cmake. String values
are used as is, while integers are converted to strings and booleans are
converted to `"ON"` and `"OFF"`. This presents the following benefits:
- Removal of "-D" visual noise everywhere for ease of reading.
- The `=` operater is literal, allowing extra whitespace for more ease
  of reading.
- CMake options are now first class Nix values that can be operated on
  directly with the full strength of the Nix language. For example,
  boolean values can be used directly because the serialization logic is
  consolidated into `mkDerivation`, getting rid of helpers like `edf`
  and making it easier to re-use feature flags like `pythonSupport`.
- Sets automatically sort options, improving cache re-use.
- An override which updates a CMake flag option will only incur
  (amortized) O(1) cost with a set, as opposed to O(n) cost with an
  (unsorted) list, which would require a full linear search. It's also
  shorter to write: just use `//`.

A few other CMake flag types are supported:
- Setting a value to null will cause its key to be unset via "-U<name>".
- The `generator` key is mapped to the "-G<value>" option.
- The `extraArgs` key is an escape hatch; any flags in this list of
  strings are not pre-processed, but concatened to the generated list
  and under-go tab escaping and joining.

`overrideAttrs` still needs to be used instead of `overrideDerivation`.
In fact, passing a set for `cmakeFlags` to `overrideDerivation` won't
work, because Nix does not know how to transform it into an environment
variable.