WIP: grub at feature parity with systemd-boot.
This replaces systemd-boot with grub, it is at feature parity, as in it
can do everything systemd-boot did in the previous commit.

This is the stepping stone to getting a themed grub; I first needed to
make it work and "install" it to the image. Turns out you can't use
grub-install.