gerenate correct dependencies of neovim
maintainers/scripts/update-luarocks-packages.sh > temp.nix
mv temp.nix pkgs/top-level/lua-generated-packages.nix

In order to run neovim tests, I switched on globally the compat flag of
lua5.2. I tried just overriding luaPackages.lua instead but it wouldn't
work.

running nix-shell -A neovim-local ~/nixpkgs --show-trace and
make functionaltest (you might need to run `export
	LUA_PATH="./?.lua;$LUA_PATH"` beforehand )

Results

[==========] 2734 tests from 278 test files ran. (273650.51 ms total)
[  PASSED  ] 2697 tests.

 11 SKIPPED TESTS
 6 FAILED TESTS
 26 ERRORS

Some of the errors might be due to lua incompatibilities but also the
tests might need some patching for nixos:

rovider/pythonx: Could not load Python 3:
python3 not found in search path or not executable.
python3.7 not found in search path or not executable.
python3.6 not found in search path or not executable.
python3.5 not found in search path or not executable.
python3.4 not found in search path or not executable.
python3.3 not found in search path or not executable.
python not found in search path or not executable.